nnimanager test passed, fixed bug of maxtrialnumber
