+++ ts/nni_manager/core/tuner_command_channel/websocket_channel.ts
@@ -33,6 +33,8 @@ export interface WebSocketChannel {
  *  Remember to invoke ``await channel.init()`` before doing anything else.
  **/
 export function getWebSocketChannel(): WebSocketChannel {
+    // return channelSingleton;
+    channelSingleton = new WebSocketChannelImpl();
     return channelSingleton;
 }
 
@@ -190,7 +192,8 @@ class WebSocketChannelImpl implements WebSocketChannel {
     }
 }
 
-const channelSingleton: WebSocketChannelImpl = new WebSocketChannelImpl();
+// const channelSingleton: WebSocketChannelImpl = new WebSocketChannelImpl();
+let channelSingleton: WebSocketChannelImpl;
 
 let heartbeatInterval: number = 5000;
